Add HTTPS support
See https://tickets.metabrainz.org/browse/MBH-363 - mb.org is at some
point going to move to HTTPS (even for the webservice). Without calling
`ne_ssl_trust_default_ca`, connections to caa.org fail because the
certificate issuer is not trusted.

With this patch, `ne_ssl_trust_default_ca` is called whenever the scheme
used to create the `ne_session` object is "https". This is the case if
the port used is 443.

Unconditionally calling `ne_ssl_trust_default_ca` does not work because
its implementation relies on an ssl context being set up on the session,
which is only done by libneon if the scheme is "https". If
`ne_ssl_trust_default_ca` is called on an `ne_session` without an ssl
context, a segfault happens.

The condition is hardcoded to port 443 at the moment because I can't
test anything else. Also, a quick look at `ne_session_create` showed
that it internally sets a port to 443 as well, so I can't be sure that
it would really work with ports other than 443.